+++ src/steps/iam/constants.ts
@@ -1,3 +1,8 @@
 export const STEP_IAM_ROLES = 'fetch-iam-roles';
 export const IAM_ROLE_ENTITY_CLASS = 'AccessRole';
 export const IAM_ROLE_ENTITY_TYPE = 'google_cloud_iam_role';
+
+export const STEP_IAM_SERVICE_ACCOUNTS = 'fetch-iam-service-accounts';
+export const IAM_SERVICE_ACCOUNT_ENTITY_CLASS = 'User';
+export const IAM_SERVICE_ACCOUNT_ENTITY_TYPE =
+  'google_cloud_iam_service_account';