+++ src/steps/iam/converters.ts
@@ -1,6 +1,11 @@
 import { iam_v1 } from 'googleapis';
 import { createIntegrationEntity } from '@jupiterone/integration-sdk-core';
-import { IAM_ROLE_ENTITY_CLASS, IAM_ROLE_ENTITY_TYPE } from './constants';
+import {
+  IAM_ROLE_ENTITY_CLASS,
+  IAM_ROLE_ENTITY_TYPE,
+  IAM_SERVICE_ACCOUNT_ENTITY_CLASS,
+  IAM_SERVICE_ACCOUNT_ENTITY_TYPE,
+} from './constants';
 import { generateEntityKey } from '../../utils/generateKeys';
 
 export function createIamRoleEntity(
@@ -33,6 +38,32 @@ export function createIamRoleEntity(
         custom: custom === true,
         deleted: data.deleted === true,
         permissions: data.includedPermissions,
+        etag: data.etag,
+      },
+    },
+  });
+}
+
+export function createIamServiceAccount(data: iam_v1.Schema$ServiceAccount) {
+  return createIntegrationEntity({
+    entityData: {
+      source: data,
+      assign: {
+        _class: IAM_SERVICE_ACCOUNT_ENTITY_CLASS,
+        _type: IAM_SERVICE_ACCOUNT_ENTITY_TYPE,
+        _key: generateEntityKey({
+          type: IAM_SERVICE_ACCOUNT_ENTITY_TYPE,
+          id: data.name as string,
+        }),
+        name: data.name,
+        displayName: (data.displayName || data.name) as string,
+        id: data.uniqueId as string,
+        projectId: data.projectId,
+        oauth2ClientId: data.oauth2ClientId,
+        username: data.email,
+        email: data.email,
+        enabled: data.disabled !== true,
+        description: data.description,
         etag: data.etag,
       },
     },