+++ src/steps/iam/index.test.ts
@@ -4,7 +4,7 @@ import {
 } from '@jupiterone/integration-sdk-testing';
 import { setupGoogleCloudRecording } from '../../../test/recording';
 import { IntegrationConfig } from '../../types';
-import { fetchIamRoles } from '.';
+import { fetchIamRoles, fetchIamServiceAccounts } from '.';
 import { integrationConfig } from '../../../test/config';
 
 describe('#fetchIamRoles', () => {
@@ -54,6 +54,56 @@ describe('#fetchIamRoles', () => {
             type: 'array',
             items: { type: 'string' },
           },
+          etag: { type: 'string' },
+        },
+      },
+    });
+  });
+});
+
+describe('#fetchIamServiceAccounts', () => {
+  let recording: Recording;
+
+  beforeEach(() => {
+    recording = setupGoogleCloudRecording({
+      directory: __dirname,
+      name: 'fetchIamServiceAccounts',
+    });
+  });
+
+  afterEach(async () => {
+    await recording.stop();
+  });
+
+  test('should collect data', async () => {
+    const context = createMockStepExecutionContext<IntegrationConfig>({
+      instanceConfig: integrationConfig,
+    });
+
+    await fetchIamServiceAccounts(context);
+
+    expect({
+      numCollectedEntities: context.jobState.collectedEntities.length,
+      numCollectedRelationships: context.jobState.collectedRelationships.length,
+      collectedEntities: context.jobState.collectedEntities,
+      collectedRelationships: context.jobState.collectedRelationships,
+      encounteredTypes: context.jobState.encounteredTypes,
+    }).toMatchSnapshot();
+
+    expect(context.jobState.collectedEntities).toMatchGraphObjectSchema({
+      _class: ['User'],
+      schema: {
+        additionalProperties: false,
+        properties: {
+          _type: { const: 'google_cloud_iam_service_account' },
+          _rawData: {
+            type: 'array',
+            items: { type: 'object' },
+          },
+          enabled: { type: 'boolean' },
+          projectId: { type: 'string' },
+          id: { type: 'string' },
+          oauth2ClientId: { type: 'string' },
           etag: { type: 'string' },
         },
       },